python3-requirements.txt
PyYAML
paho-mqtt
gpiozero
gevent
